requirements.txt
streamlit==1.22.0
fastapi==0.95.2
uvicorn==0.23.2
google-generativeai==0.3.0
pandas==2.1.0
scikit-learn==1.3.0
numpy==1.24.3